v1.1.0 — Nine new covariance estimators, Recipes layer, runnable examples, real benchmarks
### Added
- Boutquin.Trading.Recipes: IBacktestDataset, BacktestDatasetBuilder, FakeBacktestDataset
- Nine new covariance estimators completing the four-tier lineup (13 total):
  LedoitWolfConstantCorrelation, LedoitWolfSingleFactor, OracleApproximatingShrinkage,
  QuadraticInverseShrinkage, TracyWidomDenoised, Poet, Nercov, DoublySparse, Detoned
- docs/examples/: six runnable example docs (buy-and-hold through analytics)
- src/Examples/: --example <id> runner covering all five doc examples on synthetic data
- Benchmarks: BuyAndHoldBenchmark, MinimumVarianceBenchmark, CovarianceEstimatorBenchmarks,
  PortfolioConstructionBenchmarks replacing the toy Sharpe-ratio micro-benchmark

### Changed
- CorrelationAnalyzer, MonteCarloSimulator, covariance estimators delegate to Boutquin.Numerics
- MinimumVarianceConstruction and MeanVarianceConstruction use ActiveSetQpSolver (fixes KKT bug)
- Boutquin.Numerics bumped to 1.1.0; .slnx replaces legacy .sln

### Fixed
- MinimumVarianceConstruction constrained QP now finds the true global minimum
- Collinear factor inputs to FactorRegressor now throw CalculationException

### Removed
- Asset, eight market-data enums, FamaFrenchDataset enum, four market-data exceptions,
  Domain.Data.MarketData/FxRateData records, CholeskyQpSolver, Domain RollingWindow<T>
  (all superseded by Boutquin.MarketData.Abstractions or Boutquin.Numerics)

+++ src/Application/SlippageModels/VolumeShareSlippage.cs
@@ -91,9 +91,9 @@ public sealed class VolumeShareSlippage : ISlippageModel
             return theoreticalPrice;
         }
 
-        // Cap quantity at volume limit
-        var maxQuantity = (int)(barVolume * _volumeLimit);
-        var effectiveQuantity = Math.Min(quantity, Math.Max(maxQuantity, 1));
+        // Cap quantity at volume limit (use long to avoid overflow for high-volume securities)
+        var maxQuantity = (long)(barVolume * _volumeLimit);
+        var effectiveQuantity = (int)Math.Min((long)quantity, Math.Max(maxQuantity, 1L));
 
         // Price impact proportional to fraction of volume consumed
         var volumeFraction = (decimal)effectiveQuantity / barVolume;
